+++ lib/src/fortnite_api_io/structures/shop.dart
@@ -1,12 +1,27 @@
 /// item shop response
 class ShopResponse {
+  /// result of shop response
   late bool result = false;
+
+  /// is it full shop
   late bool fullShop = false;
+
+  /// when did shop last updated
   late dynamic lastUpdate;
+
+  /// current rotations
   late dynamic currentRotation;
+
+  /// next rotation time
   late dynamic nextRotation;
+
+  /// carousel of shop
   late dynamic carousel;
+
+  /// item shop video
   late dynamic specialOfferVideo;
+
+  /// item shop
   late List<dynamic> shop;
 
   ShopResponse(dynamic res) {